+++ packages/pattern_dots/CHANGELOG.md
@@ -2,6 +2,9 @@
 ### Changed
 - pubspec description
 
+### Fixed
+- matrix error
+
 ## 0.0.3 - 2023-04-24
 ### Fixed
 - lock display error